requirements.txt
pythonping
pyyaml